requirements.txt
google-cloud-vision
opencv-python
requests
python-dotenv
